backend/requirements.txt
fastapi==0.110.3
uvicorn[standard]==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
pydub==0.25.1
soundfile==0.12.1
librosa==0.10.2.post1
ffmpeg-python==0.2.0
faster-whisper==1.0.3
pyannote.audio==3.1.1
datasets==2.19.0
jiwer==3.0.4
rouge-score==0.1.2
httpx==0.27.0
orjson==3.10.3
python-multipart==0.0.9
reportlab==4.1.0
pytest==8.2.0
pytest-asyncio==0.23.6
huggingface-hub==0.23.4
matplotlib==3.9.0
transformers==4.41.2
torch==2.3.1
